[infra] refactor to reduce main
